backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
django-cors-headers>=4.3
PyMySQL>=1.1
python-dotenv>=1.0
google-genai>=1.0
sentence-transformers>=2.7
chromadb>=0.5
selenium>=4.20
webdriver-manager>=4.0
Pillow>=10.0
